tcldot: use 'size_t' values for offsets in 'rdr_t'
This is the equivalent of 08ce396f520c550daa58bfd5ce221c32a8a773ce. Squashes 2
-Wconversion warnings.

+++ tclpkg/tcldot/tcldot.h
@@ -16,6 +16,7 @@
 #include <common/render.h>
 #include <gvc/gvc.h>
 #include <gvc/gvio.h>
+#include <stddef.h>
 #include "tclhandle.h"
 #include "../../plugin/core/tcl_context.h"
 
@@ -73,8 +74,8 @@ extern int edgecmd(ClientData clientData, Tcl_Interp * interp,
 /* rdr_t isn't exposed by cgraph/io.c */
 typedef struct {
     const char *data;
-    int len;
-    int cur;
+    size_t len;
+    size_t cur;
 } rdr_t;
 
 extern int myiodisc_afread(void* channel, char *ubuf, int n);
